dusk-monitor-bot/requirements.txt
web3==6.15.1
python-telegram-bot==20.8
python-dotenv==1.0.0
